Fix: delete method should not receive body, switch product_id to params instead + Fix: verify if product exists before try to delete it

+++ src/routes/Product/Delete/DeleteProductUseCase.ts
@@ -1,5 +1,6 @@
 import { IProductRepository } from '@/repositories/ProductRepository';
 import ImageService from '@/services/ImageService';
+import { FormattedExpressError } from '@/utils/HandleExpressError';
 
 import { IDeleteProductRequestDTO } from './DeleteProductDTO';
 
@@ -9,6 +10,15 @@ export default class DeleteProductUseCase {
   ) {}
 
   async execute(data: IDeleteProductRequestDTO) {
+
+    const { product: productExists } = await this.productRepository.findById({ id: data.product_id });
+    if(!productExists)
+      throw new FormattedExpressError({
+        error: 'Unable to DeleteProductUseCase',
+        error_code: 'PRODUCT_NOT_FOUND',
+        description: 'Produto não existe',
+        status: 404
+      });
 
     const { product } = await this.productRepository.delete({ id: data.product_id });
